requirements.txt
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
rapidfuzz>=3.7.0
tenacity>=8.2.3
